Adjust wording of start.rst#secure-cdn-access; remove duplicate example
This section doesn't really make sense in the surrounding context, as
the example in #mathjax-CDN is already using HTTPS!  The HTTPS example
in this section duplicates that and is therefore unnecessary.  The
wording has been tweaked so that the user is aware that the example
in #mathjax-CDN is already safe and no action is necessary unless they
want to use protocol-relative addresses.