Reorganise the page hierarchy
At the moment, the nav menu contains 76 items, many of which have quite
long titles. It's hard to find the page I want.

There's a balance to be struck between menu depth and easy
discoverability of information, but this is definitely too far to one
extreme!

As a first pass, I've set the :caption: option on each of the toctrees,
replacing the manual headers on the index page, so a subsection heading
appears in the nav menu and not just on the index page.

The majority of the pages were configuration options pages. I've shoved
these into subdirectories, so only "configuration options" appears at
the root. I think this is reasonable - if you're trying to look up the
configurations for a particular output Jax, the "configuration options"
header would leap out at you.

I've also moved the "upgrading mathjax" pages into a subdirectory.

The "advanced topics" pages are too disparate to gather under one
heading, so I've left them as-is. I think the "basic usage" section
could do with some more organisation though - the "TeX and LaTeX"
support page, for example, isn't basic but is a pretty comprehensive
reference for the extensions and commands supported by the TeX input.

In my opinion, there should be a clear separation between "how do I do
this?" pages and reference pages. Many of the pages are walls of quite
technical exposition, hard to scan for the key bit of information needed
to get something done. Perhaps looking at the historical queries posted
to mathjax-users would be a good place to start a "how do I do this?"
section, allowing the rest of the docs to be rewritten as much more
structured reference information.